requirements.txt
crewai>=0.1.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
paymanai 
langchain-payman-tool>=0.1.0